fix: Migrating deployment from GCP trial to GitHub Pages

+++ Makefile
@@ -19,16 +19,24 @@ run_app:
 	ls -a pages_files
 	ls -a pages_files/assets
 
-	find pages_files -exec sed -i.bak 's|_dash-component-suites|dash-actions-tutorial\\/_dash-component-suites|g' {} \;
-	find pages_files -exec sed -i.bak 's|_dash-layout|dash-actions-tutorial/_dash-layout.json|g' {} \;
-	find pages_files -exec sed -i.bak 's|_dash-dependencies|dash-actions-tutorial/_dash-dependencies.json|g' {} \;
-	find pages_files -exec sed -i.bak 's|_reload-hash|dash-actions-tutorial/_reload-hash|g' {} \;
-	find pages_files -exec sed -i.bak 's|_dash-update-component|dash-actions-tutorial/_dash-update-component|g' {} \;
-	find pages_files -exec sed -i.bak 's|assets|dash-actions-tutorial/assets|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-component-suites|dash-actions-tutorial\\/_dash-component-suites|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-layout|dash-actions-tutorial/_dash-layout.json|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-dependencies|dash-actions-tutorial/_dash-dependencies.json|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_reload-hash|dash-actions-tutorial/_reload-hash|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-update-component|dash-actions-tutorial/_dash-update-component|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|assets|dash-actions-tutorial/assets|g' {} \;
 
-	mv pages_files/_dash-layout pages_files/_dash-layout.json
-	mv pages_files/_dash-dependencies pages_files/_dash-dependencies.json
-	mv assets/* pages_files/assets/
+	if [ -f pages_files/_dash-layout ]; then
+    mv pages_files/_dash-layout pages_files/_dash-layout.json
+	fi
+
+	if [ -f pages_files/_dash-dependencies ]; then
+    mv pages_files/_dash-dependencies pages_files/_dash-dependencies.json
+	fi
+
+	rsync -a --remove-source-files assets/ pages_files/assets/
+	rm -r assets/
+
 
 	ps -C python -o pid= | xargs kill -9
